Move additional layers into their own helper file/base class
